// src/app/core/enums/material-content-type.ts
export enum MaterialContentType {
    TIPS = 'TIPS',
    PROGRAMATIC_CONTENT = 'PROGRAMATIC_CONTENT',
    GENERAL_CONTENT = 'GENERAL_CONTENT',
    MANUAL = 'MANUAL',
    EXERCISE = 'EXERCISE',
    QUIZ = 'QUIZ',
    TEST = 'TEST',
    HOMEWORK = 'HOMEWORK',
    PRACTICAL = 'PRACTICAL',
    BUSINESS = 'BUSINESS',
    OTHER = 'OTHER'
}